// frontend/src/app/api/model/loginDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AttributesDto } from './attributesDto';
import { CredentialsDto } from './credentialsDto';


export interface LoginDto { 
    username?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    emailVerified?: boolean;
    enabled?: boolean;
    credentials?: CredentialsDto;
    attributes?: AttributesDto;
}